+++ packages/eyes-sdk-core/lib/cropping/FixedCutProvider.js
@@ -9,7 +9,7 @@ class FixedCutProvider extends CutProvider {
    * @param {number} left - The left to cut in pixels.
    * @param {number} right - The right to cut in pixels.
    */
-  constructor(header, footer, left, right) {
+  constructor(header = 0, footer = 0, left = 0, right = 0) {
     super()
 
     this._header = header
@@ -55,6 +55,10 @@ class FixedCutProvider extends CutProvider {
     const scaledRight = Math.ceil(this._right * scaleRatio)
 
     return new FixedCutProvider(scaledHeader, scaledFooter, scaledLeft, scaledRight)
+  }
+
+  toObject() {
+    return {left: this._left, right: this._right, top: this._header, bottom: this._footer}
   }
 }